requirements.txt
openai-whisper
pyannote.audio
torch
torchaudio
av
librosa
audioop-lts
gradio
numpy
